requirements.txt
aniso8601==7.0.0
anyio==3.3.4
asgiref==3.4.1
cached-property==1.5.2
click==8.0.3
Django==3.2.9
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.12.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
idna==3.3
pkg_resources==0.0.0
promise==2.3
psycopg2-binary==2.9.1
Pygments==2.10.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
Rx==1.6.1
sentinel==0.3.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
text-unidecode==1.3
typing-extensions==3.10.0.2
uvicorn==0.15.0
